+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `ProductPrice`.
 - Snapshot tests for `BuyButton`.
 - Snapshot tests for `AvailabilitySubscriber`.
+- Snapshot tests for `Animation`.
 
 ## [3.17.2] - 2019-02-18